new platform and compiler identification mechanism
The use of idplatform.py in config directories to determine the
platform name and compiler name (optional) has been changed a little.
The function names are now

  def get_platform()
  def get_compiler( platname, onopts )

where the 'onopts' are strictly the -o command line options (a list).
Also, the behavior is a little different. The config directories are
visited looking for idplatform.py files and the get_platform() function
is called until a non-None (and non-empty) platform name is returned.
Only then is the search stopped. Before, the search was stopped at the
first idplatform.py found.

The get_compiler() behavior is the same as get_platform().

The advantage to this behavior is that now a project can add an
idplatform.py file without repeating all of the contents/logic for
platforms known to vvtest (the builtin logic).

Backward compatibility has been maintained, so that existing
idplatform.py files will still work.