now RESOURCES_np and RESOURCES_ndevice are always available
The vvtest_util.py and vvtest_util.sh scripts will now always define
RESOURCES_np to the number of cores, and RESOURCES_ndevice to the
number of devices (or zero if there are no devices).

The use case is when the 'nnode' parameter is used to specify the
number of cores and devices, because in that case, the 'np' and
'ndevice' parameters may not be available in the vvtest_util.py file.