remove platform_plugin.py in favor of idplatform.load_specifications()
This changes the mechanism for defining platform specifictions from the
platform_plugin.py approach to another function in the idplatform.py
file. The function is

    def load_specifications( specs, platname, cplrname, options ):
        specs['batchsys'] = 'slurm'
        specs['cores_per_node'] = 36

where 'specs' is a dict-like object that receives the specifications,
and 'options' is a list of the command line -o option values.

The previous platform_plugin.py method will still work, but is now
deprecated (and should be removed in Feb 2023 or so).