fix "depends on" when using --minimal-xdirs
The --minimal-xdirs option causes parameters with only one value to be
hidden from the execute directory and the test display string. But the
"depends on" logic was using the display string for its expression
matching, and so would fail to match those hidden parameters.

Now the matching uses a display string that includes the hidden parameters.

Issue #114