modify -S param=value behavior
Previously, using -S param=value simply replaced the parameter values
in a test with the value(s) given on the command line (a single value
in this example). However, this meant that staged tests would get a
different number of stages.

So the behavior was changed. Now, for staged tests, the original number
of stages is retained.

A notable consequence in implementing this behavior change, is that
duplicate parameter values in non-staged tests are silently removed.

Issue #117