scans now default to vvt only; fix --run-dir
Support for XML formatted tests is deprecated, but still works.
This commit changes the default scan behavior from both vvt and xml
to just vvt. To recover the old behavior, add "--scan-type vvt,xml"
to the vvtest command line.

Using the --run-dir option could result in test scans that pick up
files in the test results directory. This fixes that issue.

The --run-dir option can now be an absolute path, or a directory
with more than one path segment.

Issue #82
Closes Issue #120