skip analyze tests if all dependencies are skipped
Consider the case of an execute/analyze test that does not define
a resource parameter, such as
  #VVT: parameterize : x = 1 2 3
  #VVT: analyze : --analyze
If the command line option "-p np<=4" was used, then before this
commit, none of the execute tests would run (the dependencies) but
the analyze test would still run. After this change, the analyze
test would also be skipped.

Close issue #134